Use high accuracy transform for BNG tilesource
For reasons that aren't immediately obvious, it appears that the version of
proj.4 in use in Maply 4.3.1_beta3 has some trouble with BNG tile sources using
the 7 param transformation, resulting in a blank map.

Switching to the OSTN02 transformation when creating the coordinate
system fixes this. Further investigation required to determine the root
cause.